requirements.txt
# Core dependencies
gradio>=4.0.0
pydantic>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0

# Optional: for development
# pytest
# black
# ruff
